requirements.txt
pydantic>=2.0.0
python-dateutil>=2.8.0
requests>=2.25.0
typer>=0.9.0

# Dépendances de test
pytest>=7.0.0
tenacity>=8.0.0
rich>=13.0.0
python-slugify>=8.0.0